+++ backend/src/constants/llm.py
@@ -12,7 +12,9 @@ class ModelName(str, Enum):
     OLLAMA_DEEPSEEK_R1_8B = "ollama-deepseek-r1:8b"
     OLLAMA_DEEPSEEK_R1_14B = "ollama-deepseek-r1:14b"
     GROQ_LLAMA_3_2_90B_VISION = "groq-llama-3.2-90b-vision-preview"
-    GROQ_LLAMA_3_1_70B_TOOL_USE = "llama3-groq-70b-8192-tool-use-preview"
+    GROQ_LLAMA_3_1_70B_TOOL_USE = "groq-llama3-groq-70b-8192-tool-use-preview"
+    GEMINI_PRO_1_5 = "google-gemini-1.5-pro"
+    GEMINI_PRO_2 = "google-gemini-2-pro"
     
 MODEL_CONFIG = [
     {
@@ -158,11 +160,35 @@ MODEL_CONFIG = [
             "multimodal": True,
             "embedding": False,
         }
-    }
+    },
+    {
+        "id": ModelName.GEMINI_PRO_1_5,
+        "label": "Gemini 1.5 Pro",
+        "provider": "google",
+        "metadata": {
+            "system_message": True,
+            "reasoning": False,
+            "tool_calling": False,
+            "multimodal": True,
+            "embedding": False,
+        },
+    },
+    # {
+    #     "id": ModelName.GEMINI_PRO_2,
+    #     "label": "Gemini 2 Pro",
+    #     "provider": "google",
+    #     "metadata": {
+    #         "system_message": True,
+    #         "reasoning": False,
+    #         "tool_calling": False,
+    #         "multimodal": True,
+    #         "embedding": False,
+    #     }
+    # }
 ]
 
 def get_available_models():
-    from src.constants import OPENAI_API_KEY, ANTHROPIC_API_KEY, OLLAMA_BASE_URL, GROQ_API_KEY
+    from src.constants import OPENAI_API_KEY, ANTHROPIC_API_KEY, OLLAMA_BASE_URL, GROQ_API_KEY, GEMINI_API_KEY
     
     available_models = []
     
@@ -174,7 +200,8 @@ def get_available_models():
             (provider == "openai" and OPENAI_API_KEY) or
             (provider == "anthropic" and ANTHROPIC_API_KEY) or
             (provider == "ollama" and OLLAMA_BASE_URL) or
-            (provider == "groq" and GROQ_API_KEY)
+            (provider == "groq" and GROQ_API_KEY) or
+            (provider == "google" and GEMINI_API_KEY)
         ):
             available_models.append(model)